+++ database/factories/SubscriptionItemFactory.php
@@ -19,9 +19,9 @@ class SubscriptionItemFactory extends Factory
     /**
      * Define the model's default state.
      *
-     * @return array
+     * @return array<string, mixed>
      */
-    public function definition()
+    public function definition(): array
     {
         return [
             'subscription_id' => Subscription::factory(),